apps/backend-2/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.25.0
python-dotenv==1.0.0
openai==0.28.1
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.1.0
